bugfix: ZENKO-1522 fix CRR on copied MPU object
Change the logic to ignore the 'MPU' flag set in
replicationInfo.content array (it may not always be set), instead rely
on the object size and its number of parts.

Note: the copy code lacks integrity checking now, this would need a
rework but it's out of scope of this fix.